Testing dimensions of the maps after initialisation.
Testing dimensions of the pixel map, sub-pixel map, bias register map,
smearing map, and flatfield map immediately after initialisation.  

Currently the extension of the sub-pixel map with edge pixels is not
implemented.  Once this is in place, the test harness must be updated
(added TODO where the code of the test harness must be updated then).