Multi-telescope configuration.
Script to run the Plato Simulator for a multi-telescope configuration.
The 32 telescopes are arranged in four groups of 8 telescopes.  All
telescopes in the same group have the same FOV and the lines of sight of
the four groups are offset by an angle of 9.2 degrees from the PLM
z-axis.

For each of the telescopes a simulation will be performed (i.e. a
sub-field will be modelled with the Plato Simulator).  The sub-field has
the same dimensions (in pixels for each telescope) and is always centred
on the same coordinates (raCenter, decCenter).