Tutorial for the multi-telescope configuration.
Python notebook with a (documented) tutorial of how to run simulations
for the multi-telescope configuration (i.e. four groups of 8
telescopes).

The sub-field that is modelled has a fixed size and is centred on given
sky coordinates.  For each telescopes it is determined on which CCD this
sub-field falls and (only in case the sub-field falls entirely on that
CCD) a simulation is executed.