First version of abstract base class to model temperature variations.
Similar to DriftGenerator and JitterGenerator.
This base class is used to represent temperature variations of different components (currently: FEE and CCD).  Currently we have the option to choose between (1) temperature variations stored in a file and (2) a temperature fixed at the nominal operating temperature of the component (which is stored in the configuration file).
Only the gain (for FEE and CCD) and the electronic offset (for FEE) are temperature dependent.