Exposure time ignored/exceeded when reading jitter from file.
Reported by Carsten on 14/6/2017 (via email to Joris):
The configured ExposureTime gets ignored when using a pointing time series. For example the time series is sampled with 8 Hz and I want to use 2.3 s exposure time, but PLATOSim uses 2.375 s for the exposure time instead. Please fix this, so that the configured exposure time is used.

The solution was to update the internal time and then the time step (rather than the other way around) in Camera::exposureDetector.