Number of cosmics too high #206
For the pixel map:
Cosmics hits strike the image area during exposure and readout, so here we will have to multiply with (integration time + readout time).

For the smearing map:
Cosmic hits strike the parallel over-scan (i.e. the smearing map) during readout only, so here we will have to multiply with the readout time.

For the bias map:
The pixels in the bias maps are short-lived, as they only exist in the readout register, during its serial readout. Not only is the number of cosmic hits very small here (as the serial readout happens at 2.94MHz), the effect is also confined to one row in the serial pre-scan (since no charge of the image area is shifted down from the image area).