Update of the field distortion polynomial: updated documentation.
In the past, the field distortion could be modelled with a 1D or of 2D polynomial or 3rd degree, that converted an undistored (distorted) position into the corresponding distorted (undistorted) position.  Now we have replaced this by the 1D polynomial put forward in PLATO-DLR-PL-TN-016, but with the fitting parameters we obtained ourselves.