GitHub issue #283: Wrong number of cosmics for small images
A value of 0.07128 hits per single exposure can be interpreted as ~7 hits per 100 exposures (on average). So for a number numHits < 1.0, I would recommend to generate a uniform number u between 0 and 1. If u < numHits then numHits = 1, otherwise numHits = 0. Such a procedure guarantees the correct average of cosmic hits, as well as their random nature.