Bias register map expressed as additional columns.
Previously, the bias register map was implemented as "additional" rows (at the start of each exposure), whereas it should in fact be additional pixels at the start of each row.  The configuration parameter "NumBiasPrescanRows" has now been replaced by "NumBiasPrescanColumns" and the code has been changed accordingly.  This also means that all configuration files (input files and test data) have to be updated, as well as the tutorials and documentation.