GitHub issue #285: Implement partial readout
Readout time split into:
* duration of the readout that takes place before the next exposure can start,
* duration of the readout that takes place during the next exposure.

This depends on the camera type (normal / fast) and readout mode (nominal / partial readout).

We calculate these times in the Simulation class, from where it is passed to the detector, where it is needed to advance the internal time and is needed to account for the different noise features over the correct amount of time.