Added spectral dependency based on star surface Temperature
This currently only works for MappedPSF.

An additional line in the catalog, containing the stellar Temperature, is used to generat a black body spectrum.
PlatoSim is run in multiple different (user specified) wavelength bins, using the correct amount of Flux and Transmission Efficiency / QE in each bin.
The PSF is also read in multiple colors and convolved accordingly - A multicolor PSF for single bin (old use case) has always to be provided.

Additionally, the Subfield is split into multiple, overlapping smaller fields. This allows to use a different PSF based on the region on the CCD/SkyField.

This Commit has some remaining Issues:
- Gaussian PSF source files not yet changed, so compile error
- Some time dependent variables might not work proprly yet (e.g. Jitter works, but variable focal length not)
- Smearing BG value outside of subfield not ideal yet
- Possible things missed to be checked