Include originators of ghosts in getAllStarIDs().
The getAllStarIDs() method used to return the IDs of the stars that fall on the sub-field only.  Now also the stars that produce a ghost on the sub-field are included.

Note that some stars may fall on the sub-field and will also produce an extended ghost in the sub-field.  In that case, the star ID will occur only once in the list.