Insert the ghosts in the sub-field.
When exposing the detector, not only the stars are inserted but also the ghosts.

Note that the extended source will not be convolved with the PSF, in case of an analytical PSF, but given the size of the extended source (> 200 pixels in diameter), the influence of the PSF is negligible.  In case of mapped PSFs, the extended ghosts will be convolved with the PSF.

The symmetric point-like ghosts will be treated as regular (albeit much dimmer) stars and will therefore be convolved with the PSF.