Added method to insert extended ghosts.
The addExtendedGhost() method inserts an extended ghost (with given size and flux) at the given focal-plane coordinates.

In case of an analytical PSF, the extended ghosts will not be convolved with the PSF, for practical reasons.  Given the size of the extended ghosts (> 200 pixels in diameter), the influence of the PSF is negligible.

For the mapped PSF, the convolution with the PSF happens only after all fluxes have been inserted, so the extended ghosts will be convolved with the PSF, too.