Decrease in flux due to reflections off the CCD.
The flux of a star received at the detector decreases by 0.08% + 0.00003%, due to reflections off the CCD surface.  Note that this is also the case for stars beyond 8° from the OA.  In despite of the fact that these stars do not produce point-like ghosts, there is in fact a reflection off the CCD surface, but the reflected rays will not make it through the pupil around L3 (and will therefore not reach the entrance window, where they would be reflected back towards the CCD, producing point-like ghosts).