Fixes the implementation of the JitterFromRedNoise
Fixes the issue #508.
The implementation of the Jitter in the simulation has been changed so
that jitter that is felt on different CCDs but at the same time, is
the same (provided the simulations use the same jitter seed).

Added two extra tests to the tests/validation folder that test
this behaviour.

Changed the getYawPitchRoll function in python/platosim/simfile.py
module so that it now accepts an optional parameter getTime. If this is set
to True, the function returns the yaw, pitch, roll and time. If
False (default value) the function doesn't return the time.