Apply the BFE after flooring
Fixes issue #584. The current implementation of the BFE for PlatoSim
is not accurate for large pixels (as is the case for the Plato
mission). Because of this approximation, this implementation allows for more
charge to be distracted from a pixel than is available for very bright
pixels. This commit fixes this by first applying blooming so that
there won't be any pixels with too large values.