Added a method "setSubfieldAroundPixelRows" to simulation.py
This is a method that would mimic the behaviour of the existing method setSubfieldAroundPixelCoordinates() but only for an input row and the number of lines of the subfield. This would be useful to simulate partial readout where the subfield would consist out of a few hundred lines, but all the 4510 rows.

Another difference with setSubfieldAroundPixelCoordinates() is that if it will not be possible to centre the subfield around the input row (because it is too close to the edge with respect to the number of lines of the subfield), the centre of the subfield will be shifted so that the subfield still falls onto the CCD with the correct number of lines.