F-Cam implementation
Fixes issue #605. Changed the selection of the subfield when dealing
with Fast-Cameras in `Detector::configure`. When dealing with Fast-Cameras, only the part of the subfield that is
exposed (and not shielded off) is taken into account. When no part of
the subfield is exposed an error is generated and the simulation is
terminated.