Implemented mapped field distortion for detectors with Mapped PSF
For detectors with analytic PSFs, the previous radial-symmetric
distortion model is still applied. Distortion can still be switched
on/off for such detectors.

For detectors with mapped PSFs, we used a mapped distortion. The
conversion tables for this mapping should be included in the psf file. This
behaviour can not be switched off.