Made changes to stellar aberration test
Added two more tests that compares the simulation with and without
aberration for different stellar positions.

For the first test, we assume a stellar position parallel to the
velocity and expect no/very small aberration shift.

For the second test, we assume an orthogonal stellar position to the
velocity and expect a large aberration shift.