Added new starcatalog
Added a new starcatalog in the inputfiles that more closely reflects
the PIC catalogue 1.1.0.

The pointing of the spacecraft has another default value that is
consistent with the new starcatalog.

We also included the old starcatalog for users that like to keep using
this in the near future. Remark that in order for this to work, the
pointing should be reset in the input file to (180, -70).