Changed how CCD pixel coordinates are rounded
In the function calculateSubfieldAroundCoordiantes of
referenceFrames.py the pixel CCD coordinates are converted to int
using the int() function instead of the previously used round()
function. Because of this we are sure this values is rounded down to
the lower int instead of the nearest int.

We also change in simulation.py the method
setSubfieldAroundCoordinates how the zeroPointRow/Column is set so
that the target is always in the center of the subfield.