Added estimate for initial value of occupied traps (CTI)
Previously initial number of occupied traps (inot) needed for determining the
CTI (Short2013) was set to 0. Because of the initial charge loss due
to CTI was too high in the pixelmap. We implimented a proper estimate
for the inot as described in Short2013 and proper bookeeping of the
number of occupied traps seen in pixelmap/smearing map.

This solved Issue #692.