Implemented python code for new PSF-HDF5 structure +
some bug fixes in the c++ code dealing with this new structure.

-> The major changes are in the functions that distort and undistort
FP coordintes. If no coordinates table is given in the HDF5-PSF file,
these function will try to create such a table.
REMARK: If this function is used with such a HDF5-PSF file, one should
also include the focal length as extra argument.

-> In the c++ code there was:
- One bug in HDF5File.cpp when trying to read an array attribute in
the function readArrayDatasetAttribute.
- One bug in PointSpreadFunction.cpp where we did not use correct
focalLengthValue when focalLengthSource was ConstantValue.