Depricated the use of the normal keyword in skyToPixelCoordinates...
... in referenceFrames.py.

This value will from now on be infered form the simulation object that
is given as argument.

It is (for now) still possible to use the keyword, but a warning will
be generated to the user. If the keyword does not match the value of
the Simulation object another warning will be given and the method
will use the value as given from the normal keyword.